requirements.txt
discord.py
prettytable 